requirements.txt
absl-py==0.5.0
astor==0.7.1
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
numpy==1.15.2
progressbar2==3.38.0
protobuf==3.6.1
python-utils==2.3.0
PyYAML==3.13
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
tensorboard==1.11.0
tensorflow-gpu==1.11.0
termcolor==1.1.0
Werkzeug==0.14.1
